# apps/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/settlements": {
      "post": {
        "operationId": "SettlementController_createSettlement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSettlementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "POST /settlements — 정산 생성 (관리자 전용)",
        "tags": [
          "Settlement"
        ]
      },
      "get": {
        "operationId": "SettlementController_listMySettlements",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SettlementResponse"
                  }
                }
              }
            }
          }
        },
        "summary": "GET /settlements — 판매자 본인 정산 내역 (본인만)",
        "tags": [
          "Settlement"
        ]
      }
    },
...
  },
  "info": {
    "title": "DOA Market API",
    "description": "DOA Market 백엔드 HTTP 계약 (프론트 코드젠 SSOT).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateSettlementDto": {
        "type": "object",
        "properties": {
          "sellerId": {
            "type": "string",
            "description": "정산 대상 판매자 ID (users.sellers.id)"
          },
          "periodStart": {
            "type": "string",
            "description": "정산 기간 시작 (ISO 8601)"
          },
          "periodEnd": {
            "type": "string",
            "description": "정산 기간 종료 (ISO 8601)"
          }
        },
        "required": [
          "sellerId",
          "periodStart",
          "periodEnd"
        ]
      },
      "SettlementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sellerId": {
            "type": "string",
            "description": "cross-schema plain String — users.sellers.id (P-001)"
          },
          "periodStart": {
            "type": "string",
            "format": "date-time"
          },
          "periodEnd": {
            "type": "string",
            "format": "date-time"
          },
          "totalSales": {
            "type": "string",
            "example": "1000000.00",
            "description": "금전 — 정산 기간 총 매출 (P-005)"
          },
          "commission": {
            "type": "string",
            "example": "100000.00",
            "description": "금전 — 플랫폼 수수료 (P-005)"
          },
          "payoutAmount": {
            "type": "string",
            "example": "900000.00",
            "description": "금전 — 실 지급액 (P-005)"
          },
          "status": {
            "enum": [
              "pending",
              "completed"
            ],
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sellerId",
          "periodStart",
          "periodEnd",
          "totalSales",
          "commission",
          "payoutAmount",
          "status",
          "createdAt"
        ]
      },
...
    }
  }
}